backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wework-backend"
version = "0.1.0"
description = "WeWork API - cadastro de profissionais"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.41",
  "asyncpg>=0.29.0",
  "apscheduler>=3.10.4",
  "python-json-logger>=2.0.7",
  "alembic>=1.13.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.3.0",
  "email-validator>=2.2.0",
  "pyyaml>=6.0.1",
  "python-dotenv>=1.0.1"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.0",
  "anyio>=4.4.0",
  "httpx>=0.27.0",
  "aiosqlite>=0.20.0",
  "testcontainers[postgres]>=4.8.2",
  "ruff>=0.5.0",
  "black>=24.4.0",
  "isort>=5.13.0"
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = ["alembic", "build"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["B008"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["app*"]
